getting the view more independent
