+++ PhysicsWorld/PhysicsWorld/Presenters/View/PhysicsScene/View/PhysicsScene.swift
@@ -9,7 +9,7 @@ import UIKit
 import SwiftUI
 import simd
 
-class PhysicsScene: UIViewController, ObjectViewProtocol {
+class PhysicsScene: UIViewController, ViewProtocol {
     var gameTimer: Timer?
     var qtdNodes: Int
     let qtdNodesLabel = UILabel()
@@ -26,7 +26,7 @@ class PhysicsScene: UIViewController, ObjectViewProtocol {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        view.backgroundColor = .clear
+        view.backgroundColor = .black
         setupViewCode()
         startUpdateLoop()
     }
@@ -35,40 +35,28 @@ class PhysicsScene: UIViewController, ObjectViewProtocol {
         if let touch = touches.first {
             let position = touch.location(in: view)
             
-            if var touchedObject = objects.first(where: { $0.frame.contains(position) }) {
-                
-                touchedObject.forceApplyedByEnviroment = .init(x: 0, y:0)
-                touchedObject.direction = .init(x: 0, y:0)
-                
-                return
-            }
+//            if let touchedObject = objects.first(where: { $0.frame.contains(position) }) {
+//                
+//                touchedObject.forceApplyedByEnviroment = .init(x: 0, y:0)
+//                touchedObject.direction = .init(x: 0, y:0)
+//                
+//                return
+//            }
             
             let mass = Float.random(in: 1_000...50_000)
             let radius = mass / 5000
+            
             self.addObject(type: CircleView.self, position: .init(x: Float(position.x), y: Float(position.y)), radius: radius, mass: mass, in: self)
 
         }
     }
     
     func update(_ currentTime: TimeInterval){
-        for var n: ObjConformation in objects {
-            
-            if CGFloat(n.position.x) > view.bounds.maxX + view.layer.position.x || CGFloat(n.position.x) < view.bounds.minX - view.layer.position.x || CGFloat(n.position.y) > view.bounds.maxY || CGFloat(n.position.y) < view.bounds.minY {
-                
-                n.removeFromSuperview()
-                objects.removeAll(where: { $0 === n })
-                
-            }else{
-                n.update(deltatime: currentTime)
-                n.findOtherGravityForce(for: &objects, in: &n, deltaTime: deltaTime)
-            }
-        }
+        updateObjects(currentTime)
         
         self.qtdNodes = self.view.subviews.count
         qtdNodesLabel.text = "qtd nodes: \(qtdNodes)"
     }
-    
-
 }
 
 #Preview {
